Resembling embeddings added
